Moved participant_factory to own file
