+++ backend/users/participant.py
@@ -293,67 +293,3 @@ class Participant(User):
         if not os.path.isdir(record_directory_path):
             os.mkdir(record_directory_path)
         return record_directory_path + "/" + self.id
-
-
-async def participant_factory(
-    offer: RTCSessionDescription,
-    participant_id: str,
-    experiment: _exp.Experiment,
-    participant_data: ParticipantData,
-    config: Config,
-) -> tuple[RTCSessionDescription, Participant]:
-    """Instantiate connection with a new Participant based on WebRTC `offer`.
-
-    Instantiate new modules.participant.Participant, handle offer using
-    modules.connection.connection_factory and set connection for the Participant.
-
-    This sequence must be donne for all participants.  Instantiating a Participant
-    directly will likely lead to problems, since it won't have a Connection.
-
-    Parameters
-    ----------
-    participant_data : ParticipantData
-    offer : aiortc.RTCSessionDescription
-        WebRTC offer for building the connection to the client.
-    participant_id : str
-        Unique identifier for Participant.  Must exist in experiment.
-    experiment : modules.experiment.Experiment
-        Experiment the participant is part of.
-    config : modules.config.Config
-        Hub configuration / Config object.
-
-    Returns
-    -------
-    tuple with aiortc.RTCSessionDescription, modules.participant.Participant
-        WebRTC answer that should be send back to the client and Participant
-        representing the client.
-    """
-    participant = Participant(participant_id, experiment, participant_data)
-    filter_api = FilterAPI(participant)
-    record_data = (experiment.session.record, participant.get_recording_path())
-    log_name_suffix = f"P-{participant_id}"
-
-    if config.participant_multiprocessing:
-        answer, connection = await connection_subprocess_factory(
-            offer,
-            participant.handle_message,
-            log_name_suffix,
-            config,
-            participant_data.audio_filters,
-            participant_data.video_filters,
-            filter_api,
-            record_data
-        )
-    else:
-        answer, connection = await connection_factory(
-            offer,
-            participant.handle_message,
-            log_name_suffix,
-            participant_data.audio_filters,
-            participant_data.video_filters,
-            filter_api,
-            record_data
-        )
-
-    participant.set_connection(connection)
-    return (answer, participant)
